+++ src/digitalocean/mod.rs
@@ -1,3 +1,5 @@
+//! DigitalOcean Module.
+
 pub mod apps;
 pub mod auth;
 pub mod deployments;
@@ -5,6 +7,7 @@ pub mod error;
 
 use reqwest::Client;
 
+/// Main DigitalOcean struct
 #[derive(Debug)]
 pub struct DigitalOcean {
     token: String,