+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `humantime` crate dependency
 - `thiserror` crate dependency
 
+### Changed
+
+- `digitalocean` module
+
 ## 0.1.2 - 2022-01-23
 
 ### Added